add cast manager to separate out casting logic
