+++ SlidesCast/Views/PhotoOverlay.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import SwiftUI
-import GoogleCast
 
 struct PhotoOverlay: View {
     @Binding var isShowing: Bool
@@ -23,45 +22,8 @@ struct PhotoOverlay: View {
            }
            .padding()
        } .onAppear {
-           print(imgDetails.filename)
-           if let session = GCKCastContext.sharedInstance().sessionManager.currentSession, session.connectionState == .connected {
-               saveImageToTempDirectory()
-           }
+           saveImageToTempDirectory()
        }
-    }
-
-    func castPhoto() {
-        guard let session = GCKCastContext.sharedInstance().sessionManager.currentSession else {
-            print("No Cast session is currently established.")
-            return
-        }
-        
-        guard let serverURL = LocalServer.getAddress() else {
-            print("unable to retrieve local server address")
-            return
-        }
-        guard let photoURL = URL(string: serverURL + "/images/" + imgDetails.filename) else {
-            print("Failed to create URL")
-            return
-        }
-        
-        // Create a GCKMediaMetadata object
-        let metadata = GCKMediaMetadata(metadataType: .photo)
-        
-        // Add metadata (optional, but recommended for a better user experience)
-        metadata.setString("Image Title", forKey: kGCKMetadataKeyTitle)
-        metadata.setString("A description of the image", forKey: kGCKMetadataKeySubtitle)
-        
-        // Use the builder pattern
-        let mediaInfoBuilder = GCKMediaInformationBuilder(contentURL: photoURL)
-        mediaInfoBuilder.streamType = .none // For images, use .none
-        mediaInfoBuilder.contentType = "image/jpeg" // Adjust based on your image's format
-        mediaInfoBuilder.metadata = metadata
-        
-        // Finally, build the GCKMediaInformation object
-        let mediaInformation = mediaInfoBuilder.build()
-        
-        session.remoteMediaClient?.loadMedia(mediaInformation)
     }
     
     func saveImageToTempDirectory() {
@@ -76,8 +38,9 @@ struct PhotoOverlay: View {
             do {
                 try FileManager.default.createDirectory(at: imagesDirectory, withIntermediateDirectories: true, attributes: nil)
                 FileManager.default.createFile(atPath: imagePath.path(), contents: imageData)
-                print("Image saved to: \(imagePath)")
-                castPhoto()
+                if CastManager.isCasting {
+                    CastManager.castPhoto(filename: imgDetails.filename)
+                }
             } catch {
                 print("Error saving image: \(error)")
             }
